requirements.txt
openai
requests
python-dotenv
transformers
torch
